+++ src/main/java/com/inq/wishhair/wesharewishhair/auth/controller/dto/request/MailRequest.java
@@ -1,5 +1,4 @@
-package com.inq.wishhair.wesharewishhair.user.controller.dto.request;
-
+package com.inq.wishhair.wesharewishhair.auth.controller.dto.request;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -11,8 +10,8 @@ import javax.validation.constraints.NotNull;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class AuthKeyRequest {
+public class MailRequest {
 
     @NotNull
-    private String authKey;
+    private String email;
 }
