+++ src/test/java/com/inq/wishhair/wesharewishhair/auth/controller/MailAuthControllerTest.java
@@ -5,13 +5,11 @@ import com.inq.wishhair.wesharewishhair.auth.event.AuthMailSendEvent;
 import com.inq.wishhair.wesharewishhair.global.base.ControllerTest;
 import com.inq.wishhair.wesharewishhair.global.exception.ErrorCode;
 import com.inq.wishhair.wesharewishhair.global.exception.WishHairException;
-import com.inq.wishhair.wesharewishhair.user.controller.dto.request.AuthKeyRequest;
-import com.inq.wishhair.wesharewishhair.user.controller.dto.request.MailRequest;
-import org.assertj.core.api.Assertions;
+import com.inq.wishhair.wesharewishhair.auth.controller.dto.request.AuthKeyRequest;
+import com.inq.wishhair.wesharewishhair.auth.controller.dto.request.MailRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.mockito.BDDMockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
